C#: Re-factor more the feed related methods into the FeedManager.
